move test on ad to attestation_doc module
